new project structre and code refactoring
