+++ speakleash/category_manager/__init__.py
@@ -1,0 +1,6 @@
+"""
+Speakleash Category Manager module.
+
+This module provides the CategoryManager class for managing categories in Speakleash.
+"""
+from .category_manager import CategoryManager
