+++ speakleash/config.json
@@ -1,0 +1,6 @@
+{
+  "url_datasets_text_pl": "https://speakleash.space/datasets_text/",
+  "url_datasets_text_hr": "https://speakleash.space/datasets_text_hr/",
+  "url_categories_pl": "https://speakleash.space/datasets_text/categories_pl.txt",
+  "url_categories_en": "https://speakleash.space/datasets_text/categories_en.txt"
+}
