+++ speakleash/config_2.json
@@ -1,0 +1,16 @@
+{
+    "datasets_categories": {
+        "pl": "https://speakleash.space/datasets_text/categories_pl.txt",
+        "en": "https://speakleash.space/datasets_text/categories_en.txt"
+    },
+    "datasets_text": {
+        "default": {
+            "url": "https://speakleash.space/datasets_text/",
+            "structure_file": "speakleash.json"
+        },
+        "hr": {
+            "url": "https://speakleash.space/datasets_text_hr/",
+            "structure_file": "speakleash_hr.json"
+        }
+    }
+}
