+++ speakleash/core/__init__.py
@@ -1,0 +1,6 @@
+"""
+The Speakleash main module.
+
+This module provides the Speakleash class for managing and accessing datasets.
+"""
+from .core import Speakleash
