+++ speakleash/dataset/__init__.py
@@ -1,0 +1,6 @@
+"""
+Speakleash Dataset module.
+
+This module provides the SpeakleashDataset class for handling datasets in Speakleash.
+"""
+from .dataset import SpeakleashDataset
