+++ speakleash/downloader/__init__.py
@@ -1,0 +1,6 @@
+"""
+Speakleash Structure Downloader module
+
+This package provides the StructureDownloader class for downloading dataset structures in Speakleash.
+"""
+from .structure_downloader import StructureDownloader
